pyproject.toml
[project]
name = "simple-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pygame-ce>=2.5.5",
]
